pyproject.toml
[tool.poetry]
name = "tempuscator"
readme = "README.md"
authors = ["Arunas Grigalionis <arunas.grigalionis@gmail.com>"]
description = "Script to run some sqls"
license = "GPL-3.0-only"
homepage = "https://github.com/niekosau"
repository = "https://github.com/niekosau/tempuscator"
version = "0.3.0a3"

[tool.poetry.dependencies]
python = "~3.9"
sqlalchemy = "^2.0.32"
psutil = "^6.0.0"
pymysql = "^1.1.1"
sentry-sdk = "^2.13.0"
inotify = "^0.2.10"
gitpython = "^3.1.43"
paramiko = "^3.4.1"
daemonize = "^2.5.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
tox = "^4.17.1"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"

[tool.poetry.scripts]
mysql-obfuscator = 'tempuscator.cli:obfuscator'
mysql-dir-swapper = 'tempuscator.cli:swapper'
mysql-obf-watcher = 'tempuscator.cli:mysql_obf_watcher'

[tool.pytest.ini_options]
addopts = "--cov=tempuscator"
